requirements.txt
Django #>=4.0.3,<4.1
psycopg2>=2.9.3,<2.10
import-export==0.2.67.dev6
django-import-export #==2.7.1
# pandas>=1.4.1,<1.5
xlrd2>=1.3.4,<1.4
openpyxl>=3.0.9,<3.1
requests>=2.27.1,<3
dhis2.py>=2.3.0,<2.4
django-celery-beat #>=2.2.1,<2.3
django-celery-results #>=2.2.0,<2.3
redis>=3.5.3,<3.6